rename GWmodelUtil to SpatialStats
