+++ src/main/scala/whu/edu/cn/debug/SpatialStats/BasicStatistics/readme
