+++ src/main/scala/whu/edu/cn/debug/SpatialStats/STCorrelations/TemporalAutoCorrelation.scala
@@ -1,4 +1,4 @@
-package whu.edu.cn.debug.GWmodelUtil.STCorrelations
+package whu.edu.cn.debug.SpatialStats.STCorrelations
 
 import breeze.linalg.{DenseVector, linspace}
 import breeze.plot._
