+++ src/main/scala/whu/edu/cn/debug/SpatialStats/Utils/FeatureDistance.scala
@@ -1,4 +1,4 @@
-package whu.edu.cn.debug.GWmodelUtil.Utils
+package whu.edu.cn.debug.SpatialStats.Utils
 
 import org.apache.spark.SparkContext
 import breeze.linalg.DenseMatrix
